# MemoryDiagnoser fix