--- a/src/BenchmarkDotNet/Engines/GcStats.cs
+++ b/src/BenchmarkDotNet/Engines/GcStats.cs
@@ -2,6 +2,7 @@
 using System.Reflection;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Portability;
+using BenchmarkDotNet.Running;
 using JetBrains.Annotations;
 
 namespace BenchmarkDotNet.Engines
@@ -38,19 +39,15 @@ namespace BenchmarkDotNet.Engines
 
         public long TotalOperations { get; }
 
-        public long BytesAllocatedPerOperation
+        public long GetBytesAllocatedPerOperation(BenchmarkCase benchmarkCase)
         {
-            get
-            {
-                bool excludeAllocationQuantumSideEffects = !RuntimeInformation.IsNetCore
-                    || RuntimeInformation.GetCurrentRuntime().RuntimeMoniker == RuntimeMoniker.NetCoreApp20; // the issue got fixed for .NET Core 2.0+ https://github.com/dotnet/coreclr/issues/10207
+            bool excludeAllocationQuantumSideEffects = benchmarkCase.GetRuntime().RuntimeMoniker <= RuntimeMoniker.NetCoreApp20; // the issue got fixed for .NET Core 2.0+ https://github.com/dotnet/coreclr/issues/10207
 
-                return GetTotalAllocatedBytes(excludeAllocationQuantumSideEffects) == 0
-                    ? 0
-                    : (long) Math.Round( // let's round it to reduce the side effects of Allocation quantum
-                        (double) GetTotalAllocatedBytes(excludeAllocationQuantumSideEffects) / TotalOperations,
-                        MidpointRounding.ToEven);
-            }
+            return GetTotalAllocatedBytes(excludeAllocationQuantumSideEffects) == 0
+                ? 0
+                : (long) Math.Round( // let's round it to reduce the side effects of Allocation quantum
+                    (double) GetTotalAllocatedBytes(excludeAllocationQuantumSideEffects) / TotalOperations,
+                    MidpointRounding.ToEven);
         }
 
         public static GcStats operator +(GcStats left, GcStats right)