--- a/src/BenchmarkDotNet/Exporters/Csv/CsvMeasurementsExporter.cs
+++ b/src/BenchmarkDotNet/Exporters/Csv/CsvMeasurementsExporter.cs
@@ -69,7 +69,7 @@ namespace BenchmarkDotNet.Exporters.Csv
                 new MeasurementColumn("Gen_0", (_, report, __) => report.GcStats.Gen0Collections.ToString(summary.GetCultureInfo())),
                 new MeasurementColumn("Gen_1", (_, report, __) => report.GcStats.Gen1Collections.ToString(summary.GetCultureInfo())),
                 new MeasurementColumn("Gen_2", (_, report, __) => report.GcStats.Gen2Collections.ToString(summary.GetCultureInfo())),
-                new MeasurementColumn("Allocated_Bytes", (_, report, __) => report.GcStats.BytesAllocatedPerOperation.ToString(summary.GetCultureInfo()))
+                new MeasurementColumn("Allocated_Bytes", (_, report, __) => report.GcStats.GetBytesAllocatedPerOperation(report.BenchmarkCase).ToString(summary.GetCultureInfo()))
             };
 
             return columns.ToArray();