use benchmark process runtime (not host process runtime) when deciding whether allocation quantum side effects should be excluded

Repro reported offline by @stephentoub

```cs
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Buffers;
using System.Linq;
using System.IO;
using System.Threading.Tasks;
using System.Security.Cryptography;
[MemoryDiagnoser]
public class Program
{
    public static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
    private byte[] _data = Enumerable.Range(0, 10_000_000).Select(i => (byte)i).ToArray();
    private MemoryStream _destination = new MemoryStream();
    [Benchmark]
    public async Task Encode()
    {
        _destination.Position = 0;
        using (var toBase64 = new ToBase64Transform())
        using (var stream = new CryptoStream(_destination, toBase64, CryptoStreamMode.Write, leaveOpen: true))
        {
            await stream.WriteAsync(_data, 0, _data.Length);
        }
    }
}
```

```cmd
dotnet run -c Release -f net48 --filter * --runtimes net5.0 --memory
```

Before:

```ini
BenchmarkDotNet=v0.13.0.20210803-develop, OS=Windows 10.0.18363.1621 (1909/November2019Update/19H2)
Intel Xeon CPU E5-1650 v4 3.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT
  Job-YHYGQI : .NET 5.0.8 (5.0.821.31504), X64 RyuJIT

Runtime=.NET 5.0  Toolchain=net5.0
```

| Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|--------:|--------:|------:|------:|------:|----------:|
| Encode | 172.6 ms | 2.29 ms | 2.03 ms |     - |     - |     - |         - |

After:

```ini
BenchmarkDotNet=v0.13.0.20210803-develop, OS=Windows 10.0.18363.1621 (1909/November2019Update/19H2)
Intel Xeon CPU E5-1650 v4 3.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT
  Job-YHYGQI : .NET 5.0.8 (5.0.821.31504), X64 RyuJIT

Runtime=.NET 5.0  Toolchain=net5.0
```

| Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|--------:|--------:|------:|------:|------:|----------:|
| Encode | 177.1 ms | 2.17 ms | 2.03 ms |     - |     - |     - |     768 B |